# Speed up slow Kubeadm unit tests, avoid mutating global variables