--- a/cmd/kubeadm/app/cmd/phases/init/certs_test.go
+++ b/cmd/kubeadm/app/cmd/phases/init/certs_test.go
@@ -44,7 +44,7 @@ func TestCertsWithCSRs(t *testing.T) {
 	defer os.RemoveAll(csrDir)
 	certDir := testutil.SetupTempDir(t)
 	defer os.RemoveAll(certDir)
-	cert := &certs.KubeadmCertAPIServer
+	cert := certs.KubeadmCertAPIServer()
 
 	certsData := &testCertsData{
 		cfg: testutil.GetDefaultInternalConfig(t),