**What type of PR is this?**

/kind bug
/kind flake

**What this PR does / why we need it**:

1. Ensures all kubeadm unit tests have unique names
2. Runs slow upgrade tests in parallel (xref https://github.com/kubernetes/kubernetes/issues/94528#issuecomment-687208028 and https://testgrid.k8s.io/sig-release-master-blocking#bazel-test-master&include-filter-by-regex=%2Fupgrade&width=20)

    ```
    TestStaticPodControlPlane takes 73 seconds
    	parallelizing -> 22 seconds
    TestRenewCertsByComponent takes 18 seconds
    	parallelize -> 5 seconds
    ```

3. Avoids mutating global state in package variables (discovered when running tests in parallel)

    [KubeadmCert#GetConfig](https://github.com/kubernetes/kubernetes/blob/release-1.19/cmd/kubeadm/app/phases/certs/certlist.go#L50-L60) mutates the receiver, and pointers to package KubeadmCert variables were being passed around and referenced. This PR changes those to generator functions so we can be sure all uses get a pristine copy of the object.




**Does this PR introduce a user-facing change?**:
```release-note
NONE
```

/cc @neolit123