@liggitt 2020-09-04
flake on `k8s.io/legacy-cloud-providers/azure/cache` (xref https://github.com/kubernetes/kubernetes/issues/94528#issuecomment-687208028)