@neolit123 2020-09-04
> KubeadmCert#GetConfig mutates the receiver, and pointers to package KubeadmCert variables were being passed around and referenced. This PR changes those to generator functions so we can be sure all uses get a pristine copy of the object.

this seems to have passed review in 1.12.
the changes are fine, thanks for sending this!

/lgtm
/hold cancel