@liggitt 2020-09-04
bazel debian fetch error in https://prow.k8s.io/view/gcs/kubernetes-jenkins/pr-logs/pull/94535/pull-kubernetes-e2e-gce-100-performance/1301993935178043392/
/retest